Feat (UI): Functionality to show summation of expenditures in monthly expenditures page.
In this commit, the MonthlyExpenditure.cshtml and its .cs file have been modified to show the total expenditure based on year and month. New DTO classes have been added to map the data back to the frontend. Appsettings will have a minute change based on data import in sheet sync tool (not important).